requirements.txt
pytesseract~=0.3.13
pillow~=11.2.1
opencv-python~=4.11.0.86
pandas~=2.2.3
pdf2image~=1.17.0
numpy~=2.2.5
dateparser~=1.2.1